[REF] beesdoo_easy_my_coop: cleanup file names
