+++ beesdoo_easy_my_coop/models/__init__.py
@@ -1,5 +1,4 @@
-from . import coop
-from . import product
+from . import product_template
 from . import res_company
 from . import res_partner
 from . import subscription_request
