--- beesdoo_easy_my_coop/models/coop.py
@@ -1,27 +1,0 @@
-from odoo import api, models
-
-_MAJORITY = 18
-
-
-class SubscriptionRequest(models.Model):
-    _inherit = "subscription.request"
-
-    def get_eater_vals(self, partner, share_product_id):
-        eater = share_product_id.eater
-
-        if partner.is_company or (partner.birthdate_date and partner.age < _MAJORITY):
-            eater = "eater"
-
-        return {"eater": eater}
-
-    @api.multi
-    def validate_subscription_request(self):
-        self.ensure_one()
-
-        invoice = super(SubscriptionRequest, self).validate_subscription_request()[0]
-        partner = invoice.partner_id
-
-        vals = self.get_eater_vals(partner, self.share_product_id)
-        partner.write(vals)
-
-        return invoice
