+++ beesdoo_easy_my_coop/models/subscription_request.py
@@ -1,7 +1,9 @@
 # Copyright 2019 Coop IT Easy SC
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import fields, models
+from odoo import api, fields, models
+
+_MAJORITY = 18
 
 
 class SubscriptionRequest(models.Model):
@@ -22,3 +24,25 @@ class SubscriptionRequest(models.Model):
         if company.info_session_confirmation_required:
             required_fields.append("info_session_confirmed")
         return required_fields
+
+    def get_eater_vals(self, partner, share_product_id):
+        eater = share_product_id.eater
+
+        # if birthdate_date is not set, age is 0 and lower than _MAJORITY
+        # in that case, the configuration coming from the share product is kept
+        if partner.is_company or (partner.birthdate_date and partner.age < _MAJORITY):
+            eater = "eater"
+
+        return {"eater": eater}
+
+    @api.multi
+    def validate_subscription_request(self):
+        self.ensure_one()
+
+        invoice = super().validate_subscription_request()[0]
+        partner = invoice.partner_id
+
+        vals = self.get_eater_vals(partner, self.share_product_id)
+        partner.write(vals)
+
+        return invoice
